chore: turn off warn for not used vars
We have some unused vars for documentation purposes but
the emitted warnings get picked up by GitHub and shown in
the diff view in pull requests making it impossible to review.
Although we could suppress only those specific warnings, I think
the warning is not that useful anyway so that we can turn all of
them off.